requirements.txt
streamlit>=1.20.0
pandas
pdfplumber
python-docx
openpyxl
